Extracted pure Git access interface
