+++ src/Aeon/Automation/Git/Commit.php
@@ -2,8 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Aeon\Automation\GitHub;
+namespace Aeon\Automation\Git;
 
+use Aeon\Automation\GitHub\PullRequest;
+use Aeon\Automation\GitHub\PullRequests;
 use Aeon\Automation\Project;
 use Aeon\Calendar\Gregorian\DateTime;
 use Github\Client;
