+++ src/Aeon/Automation/Git/Tags.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Aeon\Automation\GitHub;
+namespace Aeon\Automation\Git;
 
 use Composer\Semver\Semver;
 use Composer\Semver\VersionParser;
